backend/requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Embedding models
sentence-transformers==3.0.1
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
openai-whisper==20231117

# Image & PDF processing
Pillow==10.3.0
pdfplumber==0.11.1
pymupdf==1.24.5
opencv-python-headless==4.10.0.82
moviepy==1.0.3

# Vector store
faiss-cpu==1.8.0
numpy==1.26.4

# Observability
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0
structlog==24.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.1
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.3.0
